Fixed issue #18200: User with only reponse export can export Survey + token (survey archive)

+++ application/controllers/admin/export.php
@@ -867,32 +867,35 @@ class export extends Survey_Common_Action
             if ($iSurveyID === false) {
                 continue;
             }
-            if (Permission::model()->hasSurveyPermission($iSurveyID, 'responses', 'export')) {
-                $archiveName                    = "";
-                $oSurvey                        = Survey::model()->findByPk($iSurveyID);
-                $aResults[$iSurveyID]['title']  = ellipsize($oSurvey->correct_relation_defaultlanguage->surveyls_title, 30);
-                $aResults[$iSurveyID]['result'] = false;
-
+            $archiveName                    = "";
+            $oSurvey                        = Survey::model()->findByPk($iSurveyID);
+            $aResults[$iSurveyID]['title']  = ellipsize($oSurvey->correct_relation_defaultlanguage->surveyls_title, 30);
+            $aResults[$iSurveyID]['result'] = false;
+            if (Permission::model()->hasSurveyPermission($iSurveyID, 'surveycontent', 'export')) {
                 // Specific to each kind of export
                 switch ($sExportType) {
                     // Export archives for active surveys
                     case 'archive':
-                        if ($oSurvey->isActive) {
-                            $archiveName = $this->_exportarchive($iSurveyID, false);
+                        if (Permission::model()->hasSurveyPermission($iSurveyID, 'responses', 'export')) {
+                            if ($oSurvey->isActive) {
+                                $archiveName = $this->_exportarchive($iSurveyID, false);
 
-                            if (is_file($archiveName)) {
-                                $aResults[$iSurveyID]['result'] = true;
-                                $aResults[$iSurveyID]['file']   = $archiveName;
-                                $bArchiveIsEmpty                = false;
-                                $archiveFile                    = $archiveName;
-                                $newArchiveFileFullName         = 'survey_archive_'.$iSurveyID.'.lsa';
-                                $this->_addToZip($zip, $archiveFile, $newArchiveFileFullName);
-                                unlink($archiveFile);
+                                if (is_file($archiveName)) {
+                                    $aResults[$iSurveyID]['result'] = true;
+                                    $aResults[$iSurveyID]['file']   = $archiveName;
+                                    $bArchiveIsEmpty                = false;
+                                    $archiveFile                    = $archiveName;
+                                    $newArchiveFileFullName         = 'survey_archive_'.$iSurveyID.'.lsa';
+                                    $this->_addToZip($zip, $archiveFile, $newArchiveFileFullName);
+                                    unlink($archiveFile);
+                                } else {
+                                    $aResults[$iSurveyID]['error'] = gT("Unknown error");
+                                }
                             } else {
-                                $aResults[$iSurveyID]['error'] = gT("Unknown error");
+                                $aResults[$iSurveyID]['error'] = gT("Not active.");
                             }
                         } else {
-                            $aResults[$iSurveyID]['error'] = gT("Not active.");
+                            $aResults[$iSurveyID]['error'] = gT("We are sorry but you don't have permissions to do this.");
                         }
                     break;
                     // Export printable archives for all selected surveys
